+++ src/routes/eventos/[slug]/+page.server.ts
@@ -66,11 +66,14 @@ export const load: PageServerLoad = async ({ params, locals }) => {
 				Payment: {
 					orderBy: {
 						date: 'desc' // Use 'asc' for ascending order
+						},
+						include: {
+							Comment: true
+						}
 					}
 				}
-			}
-		});
-	};
+			});
+		};
 
 	// Get the total money made from the event
 	const totalMoneyRaised = async () => {
@@ -201,6 +204,17 @@ export const actions: Actions = {
 		const price = Number(formData.get('price')) || 0;
 		const ticketType = formData.get('ticketType') as 'general_tickets' | 'ringside_tickets';
 		const paymentId = formData.get('paymentId');
+		const refund = formData.get('refund');
+		const change = formData.get('change');
+
+		const changeStatus = (refund: string, change: string) => {
+			const refundMoney = Boolean(refund);
+			const changeEvent = Boolean(change);
+
+			if (refundMoney) return 'refund';
+			if (changeEvent) return 'change';
+			return 'system';
+		}
 
 		try {
 			const updatePayment = await client.payment.update({
@@ -214,7 +228,10 @@ export const actions: Actions = {
 					customer_phone: phone as string,
 					price,
 					ticketAmount,
-					ticketsType: ticketType || 'Tandas'
+					payment_status: changeStatus(refund, change),
+					ticketsType: ticketType || 'Tandas',
+					refund: Boolean(refund),
+					changeEvent: Boolean(change)
 				}
 			});
 			return {